+++ src/components/EmailEditor/components/Addons.vue
@@ -110,7 +110,7 @@
 					</v-row>
 					<v-row>
 						<v-text-field
-							v-if="'behaviorDescription' in item && (item.state === 'enabled')"
+							v-if="'behaviorDescription' in item && (item.state === 'enabled') && currentComponentBehavior === 'both'"
 							dense
 							hide-details
 							outlined
@@ -137,7 +137,6 @@ export default {
 	},
 	data() {
 		return {
-			currentComponentBehavior: this.$store.state.editorConfig.settings.addons.components.behavior,
 			componentBehaviors: ["nested", "unique", "both"],
 		};
 	},
@@ -150,7 +149,10 @@ export default {
 			return Object.keys(this.addonArr).length;
 		},
 		behaviorDescription() {
-			return this.$store.state.editorConfig.settings.addons.components.behaviorDescription;
+			return this.$store.state.editorConfig.settings.addons.componentSystem.behaviorDescription;
+		},
+		currentComponentBehavior() {
+			return this.$store.state.editorConfig.settings.addons.componentSystem.behavior;
 		},
 	},
 	methods: {