+++ CHANGELOG.md
@@ -1,13 +1,44 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [1.4.0] 2022-11-08
-
+## [Unreleased]
 ### Added
 - behaviorDescription to components Addons
 - New title and paragraph elements to email editor
+- gtm plugin instead of direct gtm load
+- components and it's settings
 
-## [1.3.0] 2022-08-16
+### Changed
+- zangoDB moved to load from script tag because it has no es build import
+- node 18
+- vite instead of webpack
+- v2 api instead of v1
+- v3 json instead of v2
+- dynamic route import so that it gets separated into chunks on build
+
+
+### Security
+- dependency version updates
+
+### Removed
+- package-lock.json
+- webpack and vue cli deps
+
+## [1.5.0] - 2023-02-23
+### Added
+- acc (Adobe Campaign Classic) and Mailchimp templating language
+
+## [1.4.1] - 2023-02-01
+### Changed
+- Update @chamaileon-sdk/plugins version that fixes sdk init message send
+
+## [1.4.0] - 2023-01-31
+
+### Added
+- German (de) language option
+- `useBlockTitleAsMarker` config option to block libraries
+
+## [1.3.0] - 2022-08-16
 
 ### Added
 - Added html import plugin tab, its settings, and data flow
@@ -15,7 +46,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - html import plugin configuration
 
-## [1.2.0] 2022-08-05
+## [1.2.0] - 2022-08-05
 
 ### Added
 - updateHooks example code for every plugin
@@ -30,18 +61,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - vuetify color palette update in the sdk config
 - French lang option for language selector
 
-## [1.1.0] 2022-08-05
+## [1.1.0] - 2022-08-05
 
 ### Added
 - Folder handling in gallery foldertree
 - New gallery hooks
 
-## [1.0.2] 2022-07-06
+## [1.0.2] - 2022-07-06
 
 ### Changed
 - New chamaileon-sdk version that fixes inline plugin dimensions
 
-## [1.0.1] 2022-06-21
+## [1.0.1] - 2022-06-21
 
 ### Changed
 - plugins version